+++ react/components/EditorContainer/Sidebar/ComponentList/utils.ts
@@ -1,9 +1,11 @@
-import { flatten, partition } from 'ramda'
+import { curry, filter, flatten, partition } from 'ramda'
 
 import { SidebarComponent } from '../typings'
 
 import { getBlockRole } from '../../../../utils/blocks'
-import { NormalizedComponent, NormalizedRelativeRoot } from './typings'
+import { NormalizedComponent, TreesByRole } from './typings'
+
+const removeFalsyValues = filter(Boolean)
 
 export const getParentTreePath = (treePath: string): string => {
   const splitTreePath = treePath.split('/')
@@ -15,97 +17,130 @@ export const getParentTreePath = (treePath: string): string => {
   return splitTreePath.slice(0, splitTreePath.length - 1).join('/')
 }
 
-export const isRootComponent = (component: SidebarComponent) => {
-  const splitTreePath = component.treePath.split('/')
+export const isChild = (rootTreePath: string, childTreePath: string) => {
+  const splitRootTreePath = removeFalsyValues(rootTreePath.split('/'))
+  const splitChildTreePath = removeFalsyValues(childTreePath.split('/'))
 
   return (
-    splitTreePath.length === 2 ||
-    (splitTreePath.length === 3 && splitTreePath[2].startsWith('$'))
+    splitRootTreePath.length > 0 &&
+    splitRootTreePath.length !== splitChildTreePath.length &&
+    splitRootTreePath.every((path, id) => {
+      return path === splitChildTreePath[id]
+    })
   )
 }
 
-export const isRootComponentRelative = (rootComponent: SidebarComponent) => {
-  const splitTreePath = rootComponent.treePath.split('/')
+export const isRootComponent = (minimumTreePathSize: number) => (
+  component: SidebarComponent
+) => {
+  const splitTreePath = component.treePath.split('/')
 
   return (
-    splitTreePath[1].startsWith('$') ||
-    (splitTreePath.length === 3 && splitTreePath[2].startsWith('$'))
+    splitTreePath.length === minimumTreePathSize ||
+    (splitTreePath.length === minimumTreePathSize + 1 &&
+      splitTreePath[minimumTreePathSize].startsWith('$'))
+  )
+}
+
+const pathLength = (path: string) => path.split('/').length
+
+const mountFullTree = curry(
+  (leaves: SidebarComponent[], root: SidebarComponent): NormalizedComponent => {
+    if (leaves.length === 0) {
+      return { ...root, components: [], isSortable: false }
+    }
+
+    const { children, grandchildren } = getChildrenAndGrandchildren(
+      root.treePath,
+      leaves
+    )
+
+    return {
+      ...root,
+      components: children.map(nextRoot => {
+        const nextChildren = grandchildren.filter(node =>
+          isChild(nextRoot.treePath, node.treePath)
+        )
+        return mountFullTree(nextChildren, nextRoot)
+      }),
+      isSortable: false,
+    }
+  }
+)
+
+const getChildrenAndGrandchildren = (
+  rootTreePath: string,
+  nodes: SidebarComponent[]
+) => {
+  const descendants = nodes.filter(node => isChild(rootTreePath, node.treePath))
+
+  const minimumTreePathSize = descendants.reduce((min, currentComponent) => {
+    const treePathSize = currentComponent.treePath.split('/').length
+    return treePathSize < min ? treePathSize : min
+  }, Infinity)
+
+  return descendants.reduce<{
+    children: SidebarComponent[]
+    grandchildren: SidebarComponent[]
+  }>(
+    (acc, node) => {
+      const nodeRelation: keyof typeof acc =
+        minimumTreePathSize === pathLength(node.treePath)
+          ? 'children'
+          : 'grandchildren'
+
+      return {
+        ...acc,
+        [nodeRelation]: acc[nodeRelation].concat(node),
+      }
+    },
+    { children: [], grandchildren: [] }
   )
 }
 
 export const normalize = (components: SidebarComponent[]) => {
-  const [roots, leaves] = partition(isRootComponent, components)
-
-  const [relativeRoots, nonRelativeRoots] = partition(
-    isRootComponentRelative,
-    roots
+  const minimumTreePathSize = components.reduce(
+    (acc, { treePath }) => Math.min(acc, pathLength(treePath)),
+    Infinity
   )
 
-  const normalizedRelativeRoots = relativeRoots.reduce<NormalizedRelativeRoot>(
-    (acc, currRoot) => {
-      const splitTreePath = currRoot.treePath.split('/')
+  const [roots, leaves] = partition(
+    isRootComponent(minimumTreePathSize),
+    components
+  )
 
+  const trees = roots.map<SidebarComponent>(mountFullTree(leaves))
+
+  const treesByRole = trees.reduce<TreesByRole>(
+    (acc, tree) => {
+      const splitTreePath = tree.treePath.split('/')
       const treePathTail = splitTreePath[splitTreePath.length - 1]
 
       const blockRole = getBlockRole(treePathTail)
-
-      if (!blockRole) {
-        return acc
-      }
-
-      const currNormalizedRoot = { ...currRoot, isSortable: false }
-
-      if (blockRole === 'around' && splitTreePath.length === 3) {
-        return {
-          ...acc,
-          nestedAround: [...acc.nestedAround, currNormalizedRoot],
-        }
-      }
-
       return {
         ...acc,
-        [blockRole]: [...acc[blockRole], currNormalizedRoot],
+        [blockRole]: [...acc[blockRole], tree],
       }
     },
-    { after: [], around: [], before: [], nestedAround: [] }
+    { after: [], around: [], before: [], blocks: [] }
   )
 
-  const normalizedNonRelativeRoots = nonRelativeRoots.map(root => ({
-    ...root,
-    isSortable: false,
-  }))
+  const prependBlocksWithAround = flatten(
+    treesByRole.blocks.map(block => {
+      const aroundForBlock = treesByRole.around.find(aroundBlock => {
+        return isChild(block.treePath, aroundBlock.treePath)
+      })
+      return aroundForBlock ? [aroundForBlock, block] : block
+    })
+  )
 
-  const sortedRoots = [
-    ...normalizedRelativeRoots.before,
-    ...normalizedRelativeRoots.around,
-    ...flatten(
-      normalizedNonRelativeRoots.map(nonRelativeRoot => [
-        ...normalizedRelativeRoots.nestedAround.filter(
-          (nestedAround: NormalizedComponent) =>
-            nestedAround.treePath.startsWith(nonRelativeRoot.treePath)
-        ),
-        nonRelativeRoot,
-      ])
-    ),
-    ...normalizedRelativeRoots.after,
+  const reorderByRole = [
+    ...treesByRole.before,
+    ...prependBlocksWithAround,
+    ...treesByRole.after,
   ]
 
-  return leaves
-    .map(leaf => ({ ...leaf, isSortable: false }))
-    .reduce<NormalizedComponent[]>(
-      (acc, leaf) =>
-        acc.map(root =>
-          leaf.treePath.startsWith(root.treePath)
-            ? {
-                ...root,
-                components: root.components
-                  ? [...root.components, leaf]
-                  : [leaf],
-              }
-            : root
-        ),
-      sortedRoots
-    )
+  return reorderByRole
 }
 
 export const pureSplice = <T>(index: number, target: T[]) => [